feat: add initial configuration files and HTML templates

+++ gulimall-coupon/src/main/java/com/atguigu/gulimall/coupon/entity/SeckillSkuRelationEntity.java
@@ -23,35 +23,35 @@ public class SeckillSkuRelationEntity implements Serializable {
 	/**
 	 * id
 	 */
-@TableId
-private Long id;
+	@TableId
+	private Long id;
 	/**
 	 * 活动id
 	 */
-private Long promotionId;
+	private Long promotionId;
 	/**
 	 * 活动场次id
 	 */
-private Long promotionSessionId;
+	private Long promotionSessionId;
 	/**
 	 * 商品id
 	 */
-private Long skuId;
+	private Long skuId;
 	/**
 	 * 秒杀价格
 	 */
-private BigDecimal seckillPrice;
+	private BigDecimal seckillPrice;
 	/**
 	 * 秒杀总量
 	 */
-private BigDecimal seckillCount;
+	private Integer seckillCount;
 	/**
 	 * 每人限购数量
 	 */
-private BigDecimal seckillLimit;
+	private Integer seckillLimit;
 	/**
 	 * 排序
 	 */
-private Integer seckillSort;
+	private Integer seckillSort;
 
 }
